Hmm, this seems to be faster ;-) Anyway, Solaris won't like it just as the :: namespace. We'll see.
git-svn-id: https://irrlicht.svn.sourceforge.net/svnroot/irrlicht/trunk@899 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ include/irrMath.h
@@ -235,7 +235,7 @@ namespace core
 
 	inline f32 round ( f32 x )
 	{
-		return round(x);
+		return roundf(x);
 	}
 
 	REALINLINE void clearFPUException ()
